+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Fixed
+
+### Changed
+
+### Removed
+
+## [v2.4.6] - 2026-08-25
+
+### Added
+
 - Returned SciPy/HiGHS MIP node count, dual bound, and optimality gap in solver statistics and displayed them in the solve progress summary.
 
 ### Fixed